+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 * Add `A.vec_size/1` macro
 * Add `A.Vector.with_index/2`
-* Add `A.Vector.random/1`
+* Add `A.Vector.random/1`, `A.Vector.take_random/2` and `A.Vector.shuffle/1`
 
 ### Breaking changes